requirements.txt
python-telegram-bot[ext]>=21.0
fastapi>=0.115
uvicorn[standard]>=0.30
google-auth>=2.35
google-auth-oauthlib>=1.2
google-api-python-client>=2.149
litellm>=1.50
sqlalchemy[asyncio]>=2.0
asyncpg>=0.29
cryptography>=43.0
python-dotenv>=1.0
pydantic-settings>=2.5
httpx>=0.27
pdfplumber>=0.11
